Add "Unstable" status to Jenkins job results
Added a new status, "Unstable", to represent test runs that have not failed but also have not passed completely. This change modifies the model to include the "Unstable" status and updates the JenkinsAgent and the frontend to handle and display this new status accordingly. An unstable job will now be represented with a different color (rgb(255, 189, 0)). This change helps to enhance the clarity of test results' interpretation.